feat: implement BRL language changes - float removal, typed references, priority removal
- Remove float type from BRL (use decimal instead)
- Remove float literals (require 'd' suffix for decimal numbers)
- Add typed references: id<ComponentName> syntax for component-typed entity refs
- Support c.entity field access on typed references
- Remove priority concept from rules (fire in declaration order)
- Update semantic analyzer to track variable types and validate typed references
- Update all game BRL files to use decimal/d-suffix instead of float
- Update test suite (143 tests all passing)
- Update brl-specification.md documentation

Agent-Logs-Url: https://github.com/ercasta/blink-idle-rpg/sessions/6a41bc7a-853e-43cf-b3c5-930fe77805d9

+++ packages/blink-compiler-ts/src/compiler.test.ts
@@ -148,7 +148,7 @@ describe('End-to-End Compiler', () => {
         component Combat {
           damage: integer
           defense: integer
-          attackSpeed: float
+          attackSpeed: decimal
         }
 
         component Target {
@@ -194,7 +194,7 @@ describe('End-to-End Compiler', () => {
         rule spawn_enemies on GameStart gs {
           let indices = [0, 1, 2, 3, 4]
           for i in indices {
-            schedule [delay: 0.5 * i] SpawnEnemy { tier: 1 }
+            schedule [delay: 0.5d * i] SpawnEnemy { tier: 1 }
           }
         }
 
@@ -223,12 +223,12 @@ describe('End-to-End Compiler', () => {
         component Combat {
           damage: integer
           defense: integer
-          attackSpeed: float
+          attackSpeed: decimal
         }
         entity {
           Character { name: "Warrior" class: "Warrior" level: 1 }
           Health { current: 100 max: 100 }
-          Combat { damage: 20 defense: 10 attackSpeed: 1.0 }
+          Combat { damage: 20 defense: 10 attackSpeed: 1.0d }
           
           .selectTarget = choice(enemies: list): id {
             let target = enemies[0]
